Insert database for seeder

+++ database/migrations/2013_03_18_900000_create_combo_table.php
@@ -18,7 +18,7 @@ class CreateComboTable extends Migration
             $table->string('name_style');
             $table->string('description');
             $table->string('image');
-            $table->float('price');
+            $table->bigInteger('price');
             $table->timestamps();
         });
     }
